requirement.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.9
django-mysql==1.1.0
djangorestframework==3.4.6
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
idna==2.6
ipaddress==1.0.19
MySQL-python==1.2.5
oauthlib==2.0.6
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22